third_party/WebKit/LayoutTests/bluetooth/idl/idl-BluetoothGATTRemoteServer.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws(null, () => new BluetoothGATTRemoteServer(),
                'the constructor should not be callable with "new"');
  assert_throws(null, () => BluetoothGATTRemoteServer(),
                'the constructor should not be callable');
}, 'BluetoothGATTRemoteServer IDL test');
</script>
